kelda: fix call to cert-gen.sh
cert-gen.sh drops the pem files in the directory it was invoked from. Since we need them in the config/ssl directory, cd there first before invoking cert-gen.sh. Do everything in a subshell so as not to affect the current working directory.

Arvados-DCO-1.1-Signed-off-by: Ward Vandewege <wvandewege@veritasgenetics.com>