kelda: Improve validation of floating IP
Before, we checked whether the user had configured a floating IP by
checking if the floating IP variable was equal to the default value.
This was error-prone because it was easy for users to accidentally
replace both the variable, and the string in the if statement.

This commit changes that check to directly check whether the IP is a
valid IP. This still allows us to ensure the user has set an IP since
"<FLOATING IP>" isn't a valid IP, with the additional benefit of
catching typos in the IP, such as "8.8.8".

This commit also moves the check to the bottom of the file to separate
it from the configuration options at the top.